+++ CHANGELOG.md
@@ -6,6 +6,7 @@ as preservation of spatial moments is not implemented
 * NNLS merging improvements (store original, non-scaled velocities and positions separately)
 * Improved test coverage
 * Slight re-ordering of `pretty_print_pia` output
+* `conservative` keyword added to `merge_roulette!` merging algorithm to ensure conservation of momentum and energy
 
 ## v0.7.3
 * Fixes in particle exchange for multi-threaded computations